Moving the assign teacher to course from course list to course edit page. fix update teacher, student and course when nothing is change don't throw error.
