+++ ASPNETCore_Assignments/DTO/CourseForUpdatingDto.cs
@@ -9,5 +9,6 @@ namespace ASPNETCore_Assignments.DTO
     [MaxLength(50)]
     public string Name { get; set; }
     public string Description { get; set; }
+    public int? TeacherId { get; set; }
   }
 }
