+++ ASPNETCore_Assignments/Reository/ITeacherRepository.cs
@@ -13,11 +13,5 @@ namespace ASPNETCore_Assignments.Reository
 		Task AddTeacherAsync(TeacherForCreatingDto dto);
 		Task DeleteTeacherAsync(int teacherId);
     Task UpdateTeacherAsync(int teacherId, TeacherForUpdatingDto dto);
-
-    Task AssignTeacherToCourseAsync(int? teacherId, int courseId);
-		Task AssignStudentsToCourseAsync(int courseId, List<int> studentsId);
-		Task AssignStudentsToCourseAsync(int courseId, List<ManageStudentInCourseDto> students);
-		void RemoveStudentsFromCourse(int courseId, List<int> studentsId);
-		void RemoveStudentsFromCourse(int courseId, List<ManageStudentInCourseDto> students);
 	}
 }
