+++ ASPNETCore_Assignments/ViewModel/SelectTeacherToCourseViewModel.cs
@@ -1,0 +1,12 @@
+﻿using ASPNETCore_Assignments.Model;
+using System.Collections.Generic;
+
+namespace ASPNETCore_Assignments.ViewModel
+{
+  public class SelectTeacherToCourseViewModel
+  {
+    public int? TeacherId { get; set; }
+    public IEnumerable<Teacher> Teachers { get; set; }
+    
+  }
+}
